requirements.txt
aiohttp==3.11.16
banks==2.1.1
blis==1.2.1
catalogue==2.0.10
certifi==2025.1.31
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.11
dirtyjson==1.0.8
distro==1.9.0
duckduckgo-search
faiss-cpu==1.10.0
filetype==1.2.0
fuzzywuzzy==0.18.0
griffe==1.7.2
httpx-sse==0.4.0
iniconfig==2.1.0
marisa-trie==1.2.1
ml-dtypes==0.5.1
murmurhash==1.0.12
neo4j==5.28.1
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
packaging==23.2
platformdirs==4.3.7
pluggy==1.5.0
primp==0.14.0
pyaudio==0.2.14
pydantic==2.11.2
pydantic-core==2.33.1
pydantic-settings==2.8.1
pypdf==4.3.1
pytest==8.3.5
pyyaml==6.0.2
requests==2.32.3
rich==14.0.0
scikit-learn==1.6.1
sentence-transformers==2.6.1
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-lookups-data==1.0.5
sqlalchemy==2.0.40
srsly==2.5.1
srt==3.5.3
striprtf==0.0.26
tensorboard==2.19.0
tensorflow==2.19.0
tf-keras==2.19.0
thinc==8.3.4
threadpoolctl==3.6.0
tomli==2.2.1
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
uvicorn==0.34.0
vosk==0.3.45
wasabi==1.1.3
weasel==0.4.1
yarl==1.19.0
langchain
langchain_community
